+++ src/main/java/io/github/thatsmusic99/headsplus/api/rewards/AddGroupReward.java
@@ -12,7 +12,7 @@ public class AddGroupReward extends Reward {
 
     private final String group;
 
-    public AddGroupReward(String group, int xp) {
+    public AddGroupReward(String group, long xp) {
         super(xp);
         this.group = group;
     }
@@ -21,7 +21,7 @@ public class AddGroupReward extends Reward {
         if (!section.contains("base-value") && !section.contains("reward-value"))
             throw new IllegalStateException("Reward type ADD_GROUP for reward " + id + " must have a base-value option!");
         return new AddGroupReward(section.getString("base-value", section.getString("reward-value")),
-                section.getInteger("base-xp", 0));
+                section.getLong("base-xp", 0));
     }
 
     @Override
