+++ src/main/java/io/github/thatsmusic99/headsplus/api/rewards/RunCommandReward.java
@@ -12,7 +12,7 @@ public class RunCommandReward extends Reward {
 
     private final List<String> commands;
 
-    public RunCommandReward(int xp, List<String> commands) {
+    public RunCommandReward(long xp, List<String> commands) {
         super(xp);
         this.commands = commands;
     }
@@ -20,7 +20,7 @@ public class RunCommandReward extends Reward {
     public static RunCommandReward fromConfigSection(String id, ConfigSection section) {
         if (!section.contains("base-value") && !section.contains("reward-value"))
             throw new IllegalStateException("Reward type RUN_COMMAND for reward " + id + " must have a base-value option!");
-        return new RunCommandReward(section.getInteger("base-xp", 0),
+        return new RunCommandReward(section.getLong("base-xp", 0),
                 section.getList("base-value", section.getList("reward-value")));
     }
